+++ modules/mpi/ospray/MPIDistributedDevice.h
@@ -1,5 +1,5 @@
 // ======================================================================== //
-// Copyright 2009-2019 Intel Corporation                                    //
+// Copyright 2009-2020 Intel Corporation                                    //
 //                                                                          //
 // Licensed under the Apache License, Version 2.0 (the "License");          //
 // you may not use this file except in compliance with the License.         //
@@ -135,6 +135,8 @@ struct MPIDistributedDevice : public api::Device
   void cancel(OSPFuture) override;
 
   float getProgress(OSPFuture) override;
+
+  float getTaskDuration(OSPFuture) override;
 
   float getVariance(OSPFrameBuffer) override;