+++ modules/mpi/ospray/common/OSPWork.h
@@ -1,5 +1,5 @@
 // ======================================================================== //
-// Copyright 2009-2019 Intel Corporation                                    //
+// Copyright 2009-2020 Intel Corporation                                    //
 //                                                                          //
 // Licensed under the Apache License, Version 2.0 (the "License");          //
 // you may not use this file except in compliance with the License.         //
@@ -81,6 +81,7 @@ enum TAG
   FUTURE_WAIT,
   FUTURE_CANCEL,
   FUTURE_GET_PROGRESS,
+  FUTURE_GET_TASK_DURATION,
   FINALIZE
 };