+++ modules/mpi/ospray/render/ThreadedRenderTask.h
@@ -1,5 +1,5 @@
 // ======================================================================== //
-// Copyright 2009-2019 Intel Corporation                                    //
+// Copyright 2009-2020 Intel Corporation                                    //
 //                                                                          //
 // Licensed under the Apache License, Version 2.0 (the "License");          //
 // you may not use this file except in compliance with the License.         //
@@ -38,8 +38,11 @@ struct ThreadedRenderTask : public Future
   void cancel() override;
   float getProgress() override;
 
+  float getTaskDuration() override;
+
  private:
   Ref<FrameBuffer> fb;
+  std::atomic<float> taskDuration{0.f};
   std::atomic<bool> finished;
   std::thread thread;
 };
@@ -51,7 +54,7 @@ inline ThreadedRenderTask::ThreadedRenderTask(
     : fb(_fb), finished(false)
 {
   thread = std::thread([this, fcn]() {
-    fcn();
+    taskDuration = fcn();
     finished = true;
   });
 }
@@ -86,6 +89,11 @@ inline void ThreadedRenderTask::cancel()
 inline float ThreadedRenderTask::getProgress()
 {
   return fb->getCurrentProgress();
+}
+
+inline float ThreadedRenderTask::getTaskDuration()
+{
+    return taskDuration.load();
 }
 
 } // namespace mpi